perf(system): Measurement is not mutable now
Before the change, Measurements keeped a variable n, marking the index of the measurement that was going to give. Now calculates and index depending on the time given by KalmanIterator. `n` is not needed anymore, and, as it was the only changing variable in Measurements, being mutable is not longer necessary.